根据apple官方文档Adopt Morden Objective-C,来规范和优化现有代码。1.instancetype 2.property 3.NS_ENUM 4.减少头文件引用其他头文件，尽量使用@class前向引用 5.const 常量代替预定义宏 6.copy修饰符的应用

+++ SSFAnalysis/Expansion/Category/NSString+Tony.m
@@ -93,14 +93,14 @@ static char base64EncodingTable[64] = {
             timeString = [NSString stringWithFormat:@"00 : %lu", (unsigned long)time];
         }
     } else if (time >= 60) {
-        int minute = time/60;
+        NSUInteger minute = time/60;
         int second = time % 60;
         NSString *minuteString;
         NSString *secondString;
         if (minute < 10) {
-            minuteString = [NSString stringWithFormat:@"0%d", minute];
+            minuteString = [NSString stringWithFormat:@"0%lu", (unsigned long)minute];
         } else {
-            minuteString = [NSString stringWithFormat:@"%d", minute];
+            minuteString = [NSString stringWithFormat:@"%lu", (unsigned long)minute];
         }
         if (second < 10) {
             secondString = [NSString stringWithFormat:@"0%d", second];
